+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## master (unreleased)
 
 * [#285](https://github.com/clojure-emacs/orchard/issues/285): **BREAKING:** Remove special handling of Boot classpath.
-* [#287](https://github.com/clojure-emacs/orchard/issues/287): Inspector: ton't crash when field contains non-equiv()able value.
+* [#287](https://github.com/clojure-emacs/orchard/issues/287): Inspector: don't crash when field contains non-equiv()able value.
+* [#288](https://github.com/clojure-emacs/orchard/issues/288): Java: don't crash when loading if none of the parsers could be correctly initialized.
 
 ## 0.26.3 (2024-08-14)